requirements.txt
beautifulsoup4==4.12.2
bleach==6.0.0
defusedxml==0.7.1
easyocr==1.6.2
hatchling==1.17.0
importlib_metadata==6.6.0
ipython==8.13.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab_pygments==0.2.2
MarkupSafe==2.1.2
mistune==2.0.5
nbclient==0.7.4
nbformat==5.8.0
numpy==1.24.3
opencv_python==4.7.0.72
opencv_python_headless==4.5.4.60
packaging==23.1
pandocfilters==1.5.0
Pillow==9.5.0
Pillow==9.5.0
pyenchant==3.2.2
Pygments==2.15.1
pyppeteer==1.0.2
PyQt5==5.15.9
pytesseract==0.3.10
pytest==7.3.1
rembg==2.0.37
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
